Integrated song search match score to SQL and added prefix query

+++ Services/SearchService/SongService/SongMetadata.cs
@@ -15,7 +15,7 @@ public class SongMetadata
 
     public string Mapper { get; private set; }
 
-    public float Score { get; set; }
+    public int Score { get; set; }
 
     public static explicit operator SongMetadata(Song song) =>
         new()
